requirements.txt
pyTigerGraph>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
tiktoken>=0.5.0
groq>=0.4.0
httpx>=0.25.0
pyvis>=0.3.2
faker>=20.0.0
cryptography>=41.0.0
grpcio>=1.59.0
python-dotenv>=1.0.0
pydantic>=2.4.0
aiofiles>=23.2.0
